+++ vertx-registry/src/test/scala/com/cloudentity/tools/vertx/registry/RegistryVerticleTest.scala
@@ -4,7 +4,7 @@ import java.io.File
 import java.util.Optional
 
 import com.google.common.io.Files
-import com.cloudentity.tools.vertx.bus.ServiceClientFactory
+import com.cloudentity.tools.vertx.bus.VertxEndpointClient
 import com.cloudentity.tools.vertx.conf.ConfVerticleDeploy
 import com.cloudentity.tools.vertx.registry.RegistryVerticle.RegistryType
 import com.cloudentity.tools.vertx.scala.{FutureConversions, Futures}
@@ -25,7 +25,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/one-healthy.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath)
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }
@@ -45,7 +45,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/disabled-healthy.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath)
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }
@@ -65,7 +65,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/enabled-false-healthy.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath)
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }
@@ -85,7 +85,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/registry-in-registry.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath)
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(RegistryType("super"))) }
@@ -118,7 +118,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/one-healthy.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath, new ConfigRetrieverOptions().setScanPeriod(100))
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }
@@ -144,7 +144,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/two-healthy.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
 
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath, new ConfigRetrieverOptions().setScanPeriod(100))
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }
@@ -175,7 +175,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/verticle-config.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath)
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }
@@ -195,7 +195,7 @@ class RegistryVerticleTest() extends VertxUnitTest with FutureConversions {
     val configFile = copyDescriptorsToTempFile("src/test/resources/registry/verticle-prefix.json")
     val typ = RegistryType("test")
 
-    val client = ServiceClientFactory.make(vertx.eventBus(), classOf[RegistryService], Optional.of(typ.value))
+    val client = VertxEndpointClient.make(vertx, classOf[RegistryService], Optional.of(typ.value))
     // when
     ConfVerticleDeploy.deployFileConfVerticle(vertx, configFile.getAbsolutePath)
       .compose { _ => VertxDeploy.deploy(vertx, new RegistryVerticle(typ)) }